third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-image-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-image with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-image">
<meta name="assert" content="mask-image supports only the grammar '<mask-reference>#', where <mask-reference> is 'none | <image> | <mask-source>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-image", "auto");
test_invalid_value("mask-image", "circle()");
test_invalid_value("mask-image", "border-box");

test_invalid_value("mask-image", "none none");
</script>
</body>
</html>
